sync up edit steps
